tools/netinspector/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "netinspect"
version = "0.1.0"
description = "Azure network topology discovery, documentation, and analysis tool"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "azure-identity>=1.15.0",
    "azure-mgmt-network>=25.0.0",
    "azure-mgmt-resource>=23.0.0",
    "azure-mgmt-subscription>=3.0.0",
    "azure-mgmt-privatedns>=1.1.0",
    "networkx>=3.2",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.3.0",
]

[project.scripts]
netinspect = "netinspect.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/netinspect"]

[tool.ruff]
src = ["src"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
